Fix issue where SiteRemover caused an internal server error
This commit fixes 2 issues with the S3SiteRemover service, and adds regression test to prevent the issues in the past.

Both issues relate to issues with how the `delete-objects` API request was sent to the S3 API.

The first issue was that an empty request (i.e. a request to delete 0 objects) would cause a the S3 API to respond with a 500 which would propagate to the S3SiteRemover. This commit adds a guard clause to the top of the `deleteObjects` function to prevent this from happening.

The second issue happened because the `delete-objects` API requests does not allow for the delete of more than 1000 objects at a time. This commit changes `deleteObjects` function so that it returns a Promise that recursively deletes objects 1000 at a time. This means at most 1000 objects are deleted at a time and no deletion requests are happening simultaneously threatening to break AWS's rate limit.

Ref #990

+++ api/services/S3SiteRemover.js
@@ -1,55 +1,70 @@
-const AWS = require("aws-sdk")
-const config = require("../../config")
+const AWS = require('aws-sdk');
+const config = require('../../config');
 
-const s3Config = config.s3
+const s3Config = config.s3;
 const s3Client = new AWS.S3({
   accessKeyId: s3Config.accessKeyId,
   secretAccessKey: s3Config.secretAccessKey,
   region: s3Config.region,
-})
+});
 
-const removeSite = site => {
-  return Promise.all([
-    getObjectsWithPrefix(`site/${site.owner}/${site.repository}`),
-    getObjectsWithPrefix(`demo/${site.owner}/${site.repository}`),
-    getObjectsWithPrefix(`preview/${site.owner}/${site.repository}`),
-  ]).then(objects => {
-    objects = [].concat.apply([], objects)
-    return deleteObjects(objects)
-  })
-}
+/**
+  Deletes the array of S3 objects passed to it.
 
-const deleteObjects = objects => {
+  Since AWS limits the number of objects that can be deleted at a time, this
+  method deletes the objects 1000 at a time. It does so recursively so each
+  group of 1000 is deleted one after the other instead of simultaneously. This
+  prevents the delete requests from breaking AWS's rate limit.
+*/
+const deleteObjects = (objects) => {
+  if (objects.length === 0) {
+    return Promise.resolve();
+  }
+
+  const objectsToDeleteNow = objects.slice(0, 1000);
+  const objectsToDeleteLater = objects.slice(1000, objects.length);
+
   return new Promise((resolve, reject) => {
     s3Client.deleteObjects({
       Bucket: s3Config.bucket,
       Delete: {
-        Objects: objects.map(object => ({ Key: object }))
+        Objects: objectsToDeleteNow.map(object => ({ Key: object })),
       },
     }, (err, data) => {
       if (err) {
-        reject(err)
+        reject(err);
       } else {
-        resolve(data)
+        resolve(data);
       }
-    })
-  })
-}
+    });
+  }).then(() =>
+    deleteObjects(objectsToDeleteLater)
+  );
+};
 
-const getObjectsWithPrefix = prefix => {
-  return new Promise((resolve, reject) => {
+const getObjectsWithPrefix = prefix =>
+  new Promise((resolve, reject) => {
     s3Client.listObjects({
       Bucket: s3Config.bucket,
       Prefix: prefix,
     }, (err, data) => {
       if (err) {
-        reject(err)
+        reject(err);
       } else {
-        const keys = data.Contents.map(object => object.Key)
-        resolve(keys)
+        const keys = data.Contents.map(object => object.Key);
+        resolve(keys);
       }
-    })
-  })
-}
+    });
+  });
 
-module.exports = { removeSite }
+const removeSite = site =>
+  Promise.all([
+    getObjectsWithPrefix(`site/${site.owner}/${site.repository}`),
+    getObjectsWithPrefix(`demo/${site.owner}/${site.repository}`),
+    getObjectsWithPrefix(`preview/${site.owner}/${site.repository}`),
+  ]).then((objects) => {
+    const mergedObjects = [].concat(...objects);
+    return deleteObjects(mergedObjects);
+  });
+
+module.exports = { removeSite };
